requirements.txt
PyMuPDF==1.27.2
python-docx==1.2.0
Pillow==12.0.0
